requirements.txt
gspread==5.12.3
oauth2client==4.1.3
requests==2.31.0
python-dotenv==1.0.0
loguru==0.7.2
schedule==1.2.1
pytz==2024.1 